% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_specieslevel.r
\name{boot_specieslevel_once}
\alias{boot_specieslevel_once}
\title{Bootstrap metric at the species level once.}
\usage{
boot_specieslevel_once(data, col_lower, col_higher, index, level, start,
  step, seed, ...)
}
\arguments{
\item{data}{See \code{\link[bootstrapnet]{boot_specieslevel_n}}.}

\item{col_lower}{See \code{\link[bootstrapnet]{boot_specieslevel_n}}.}

\item{col_higher}{See \code{\link[bootstrapnet]{boot_specieslevel_n}}.}

\item{index}{See \code{\link[bootstrapnet]{boot_specieslevel_n}}.}

\item{level}{See \code{\link[bootstrapnet]{boot_specieslevel_n}}.}

\item{start}{See \code{\link[bootstrapnet]{boot_specieslevel_n}}.}

\item{step}{See \code{\link[bootstrapnet]{boot_specieslevel_n}}.}

\item{seed}{Set seed to get reproducible random results. Passed to \code{set.seed()}.}

\item{...}{See \code{\link[bootstrapnet]{boot_specieslevel_n}}.}
}
\value{
A list of 1, 2 or 4 matrices. This depends on the provided \code{index} metric
and \code{level} value.
}
\description{
You will rarely use this function alone. It was designed to be executed in
parallel by \code{\link[bootstrapnet]{boot_specieslevel_n}}. See more
details there.
}
\references{
This function is a wrapper of \code{\link[bipartite]{specieslevel}}.
}
