% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_networklevel.r
\name{boot_networklevel_once}
\alias{boot_networklevel_once}
\title{Bootstrap network level metric once.}
\usage{
boot_networklevel_once(
  data,
  col_lower,
  col_higher,
  index,
  level,
  start,
  step,
  seed,
  ...
)
}
\arguments{
\item{data}{See \code{\link[bootstrapnet]{boot_networklevel_n}}.}

\item{col_lower}{See \code{\link[bootstrapnet]{boot_networklevel_n}}.}

\item{col_higher}{See \code{\link[bootstrapnet]{boot_networklevel_n}}.}

\item{index}{See \code{\link[bootstrapnet]{boot_networklevel_n}}.}

\item{level}{See \code{\link[bootstrapnet]{boot_networklevel_n}}.}

\item{start}{See \code{\link[bootstrapnet]{boot_networklevel_n}}.}

\item{step}{See \code{\link[bootstrapnet]{boot_networklevel_n}}.}

\item{seed}{Set seed to get reproducible random results. Passed to \code{set.seed()}.}

\item{...}{See \code{\link[bootstrapnet]{boot_networklevel_n}}.}
}
\value{
Returns a data frame of one or two columns depending on the provided \code{index}
metric and \code{level} value. For example if \code{index='nestedness'} and
\code{level='both'}, then it returns a one column data frame. But in case of
\code{index='niche overlap'}, it returns a two columns data frame, first column
('niche.overlap.HL') corresponding to the higher level (e.g. insects), and
the second column ('niche.overlap.LL') for the lower level (e.g. plants).
}
\description{
You will rarely use this function alone. It was designed to be executed in
parallel by \code{\link[bootstrapnet]{boot_networklevel_n}}. See more
details there.
}
\references{
This function is a wrapper of \code{\link[bipartite]{networklevel}}.
}
